+++ beaverbot/clutter/csv_analysis/output/Acadia2010/X_coordinates.npy
