+++ beaverbot/clutter/csv_analysis/output/Acadia2010/Y_coordinates.npy
