+++ beaverbot/clutter/csv_analysis/output/Acadia2010/elevation.npy
