+++ beaverbot/clutter/csv_analysis/output/Acadia2023/X_coordinates.npy
