+++ beaverbot/clutter/csv_analysis/output/Acadia2023/Y_coordinates.npy
