+++ beaverbot/clutter/csv_analysis/output/Acadia2023/elevation.npy
