+++ beaverbot/clutter/csv_analysis/output/Acadia2023/elevation_with_vegetation.npy
