+++ beaverbot/clutter/csv_analysis/output/Mars/X_coordinates.npy
