+++ beaverbot/clutter/csv_analysis/output/Mars/Y_coordinates.npy
