+++ beaverbot/clutter/csv_analysis/output/Mars/elevation_with_vegetation.npy
