+++ beaverbot/constants.py
