+++ beaverbot/scenarios/fedeoli_beavers_scenario.py
@@ -1,5 +1,5 @@
-from earth_moving.ral.backend.base_backend import BaseBackend
-from earth_moving.ral.environment.environment_beavers_backend import BeaversEnvironmentBackend
+from beaverbot.ral.backend.base_backend import BaseBackend
+from beaverbot.ral.environment.environment_beavers_backend import BeaversEnvironmentBackend
 from IPython.display import clear_output
 from tqdm import tqdm
